+++ scanpipe/api/views.py
@@ -247,10 +247,11 @@ class ProjectViewSet(
 
         pipeline = request.data.get("pipeline")
         if pipeline:
-            pipeline = scanpipe_app.get_new_pipeline_name(pipeline)
-            if pipeline in scanpipe_app.pipelines:
+            pipeline_name, groups = scanpipe_app.extract_group_from_pipeline(pipeline)
+            pipeline_name = scanpipe_app.get_new_pipeline_name(pipeline_name)
+            if pipeline_name in scanpipe_app.pipelines:
                 execute_now = request.data.get("execute_now")
-                project.add_pipeline(pipeline, execute_now)
+                project.add_pipeline(pipeline_name, execute_now, selected_groups=groups)
                 return Response({"status": "Pipeline added."})
 
             message = {"status": f"{pipeline} is not a valid pipeline."}