+++ scanpipe/models.py
@@ -691,7 +691,9 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, UpdateMixin, models.Model):
 
         if copy_pipelines:
             for run in self.runs.all():
-                cloned_project.add_pipeline(run.pipeline_name, execute_now)
+                cloned_project.add_pipeline(
+                    run.pipeline_name, execute_now, selected_groups=run.selected_groups
+                )
 
         if copy_subscriptions:
             for subscription in self.webhooksubscriptions.all():
@@ -1025,7 +1027,7 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, UpdateMixin, models.Model):
         for uploaded_file in uploads:
             self.add_upload(uploaded_file)
 
-    def add_pipeline(self, pipeline_name, execute_now=False):
+    def add_pipeline(self, pipeline_name, execute_now=False, selected_groups=None):
         """
         Create a new Run instance with the provided `pipeline` on the current project.
 
@@ -1039,10 +1041,13 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, UpdateMixin, models.Model):
         if not pipeline_class:
             raise ValueError(f"Unknown pipeline: {pipeline_name}")
 
+        validate_none_or_list(selected_groups)
+
         run = Run.objects.create(
             project=self,
             pipeline_name=pipeline_name,
             description=pipeline_class.get_summary(),
+            selected_groups=selected_groups,
         )
 
         # Do not start the pipeline execution, even if explicitly requested,
@@ -1585,6 +1590,11 @@ class RunQuerySet(ProjectRelatedQuerySet):
         return self.filter(task_id__isnull=False, task_end_date__isnull=True)
 
 
+def validate_none_or_list(value):
+    if value is not None and not isinstance(value, list):
+        raise ValidationError("Value must be a list.")
+
+
 class Run(UUIDPKModel, ProjectRelatedModel, AbstractTaskFieldsModel):
     """The Database representation of a pipeline execution."""
 
@@ -1596,6 +1606,9 @@ class Run(UUIDPKModel, ProjectRelatedModel, AbstractTaskFieldsModel):
     scancodeio_version = models.CharField(max_length=30, blank=True)
     description = models.TextField(blank=True)
     current_step = models.CharField(max_length=256, blank=True)
+    selected_groups = models.JSONField(
+        null=True, blank=True, validators=[validate_none_or_list]
+    )
 
     objects = RunQuerySet.as_manager()