+++ scanpipe/tests/pipelines/with_groups.py
@@ -20,37 +20,37 @@
 # ScanCode.io is a free software code scanning tool from nexB Inc. and others.
 # Visit https://github.com/nexB/scancode.io for support and download.
 
-from django.template.defaultfilters import pluralize
-
-from scanpipe.management.commands import ProjectCommand
-from scanpipe.management.commands import extract_group_from_pipelines
-from scanpipe.management.commands import validate_pipelines
+from scanpipe.pipelines import Pipeline
+from scanpipe.pipelines import group
 
 
-class Command(ProjectCommand):
-    help = "Add pipelines to a project."
+class WithGroups(Pipeline):
+    """Include "grouped" steps."""
 
-    def add_arguments(self, parser):
-        super().add_arguments(parser)
-        parser.add_argument(
-            "args",
-            metavar="PIPELINE_NAME",
-            nargs="+",
-            help="One or more pipeline names.",
+    @classmethod
+    def steps(cls):
+        return (
+            cls.grouped_with_foo_and_bar,
+            cls.grouped_with_bar,
+            cls.grouped_with_excluded,
+            cls.no_groups,
         )
 
-    def handle(self, *pipelines, **options):
-        super().handle(*pipelines, **options)
+    @group("foo", "bar")
+    def grouped_with_foo_and_bar(self):
+        """Step1 doc."""
+        pass
 
-        pipelines_data = extract_group_from_pipelines(pipelines)
-        pipelines_data = validate_pipelines(pipelines_data)
+    @group("bar")
+    def grouped_with_bar(self):
+        """Step2 doc."""
+        pass
 
-        for pipeline_name, selected_groups in pipelines_data.items():
-            self.project.add_pipeline(pipeline_name, selected_groups=selected_groups)
+    @group("excluded")
+    def grouped_with_excluded(self):
+        """Step2 doc."""
+        pass
 
-        pipeline_names = pipelines_data.keys()
-        msg = (
-            f"Pipeline{pluralize(pipeline_names)} {', '.join(pipeline_names)} "
-            f"added to the project"
-        )
-        self.stdout.write(msg, self.style.SUCCESS)
+    def no_groups(self):
+        """Step2 doc."""
+        pass