+++ scanpipe/tests/test_api.py
@@ -527,7 +527,7 @@ class ScanPipeAPITest(TransactionTestCase):
     def test_scanpipe_api_project_action_pipelines(self):
         url = reverse("project-pipelines")
         response = self.csrf_client.get(url)
-        expected = ["name", "summary", "description", "steps"]
+        expected = ["name", "summary", "description", "steps", "available_groups"]
         self.assertEqual(expected, list(response.data[0].keys()))
 
     def test_scanpipe_api_project_action_resources(self):
@@ -758,6 +758,16 @@ class ScanPipeAPITest(TransactionTestCase):
         url = reverse("project-add-pipeline", args=[self.project1.uuid])
         data = {
             "pipeline": "docker",  # old name
+            "execute_now": False,
+        }
+        response = self.csrf_client.post(url, data=data)
+        self.assertEqual({"status": "Pipeline added."}, response.data)
+        self.assertEqual("analyze_docker_image", self.project1.runs.get().pipeline_name)
+
+    def test_scanpipe_api_project_action_add_pipeline_groups(self):
+        url = reverse("project-add-pipeline", args=[self.project1.uuid])
+        data = {
+            "pipeline": "analyze_docker_image:group1,group2",
             "execute_now": False,
         }
         response = self.csrf_client.post(url, data=data)