+++ scanpipe/tests/test_commands.py
@@ -100,7 +100,7 @@ class ScanPipeManagementCommandTest(TestCase):
             "--pipeline",
             self.pipeline_name,
             "--pipeline",
-            "analyze_root_filesystem_or_vm_image",
+            "analyze_root_filesystem_or_vm_image:group1,group2",
             "--pipeline",
             "scan_package",  # old name backward compatibility
         ]
@@ -113,6 +113,8 @@ class ScanPipeManagementCommandTest(TestCase):
             "scan_single_package",
         ]
         self.assertEqual(expected, [run.pipeline_name for run in project.runs.all()])
+        run = project.runs.get(pipeline_name="analyze_root_filesystem_or_vm_image")
+        self.assertEqual(["group1", "group2"], run.selected_groups)
 
     def test_scanpipe_management_command_create_project_inputs(self):
         out = StringIO()
@@ -254,7 +256,7 @@ class ScanPipeManagementCommandTest(TestCase):
 
         pipelines = [
             self.pipeline_name,
-            "analyze_root_filesystem_or_vm_image",
+            "analyze_root_filesystem_or_vm_image:group1,group2",
             "scan_package",  # old name backward compatibility
         ]
 
@@ -276,6 +278,8 @@ class ScanPipeManagementCommandTest(TestCase):
             "scan_single_package",
         ]
         self.assertEqual(expected, [run.pipeline_name for run in project.runs.all()])
+        run = project.runs.get(pipeline_name="analyze_root_filesystem_or_vm_image")
+        self.assertEqual(["group1", "group2"], run.selected_groups)
 
         options = ["--project", project.name, "non-existing"]
         expected = "non-existing is not a valid pipeline"
@@ -301,13 +305,15 @@ class ScanPipeManagementCommandTest(TestCase):
         )
         self.assertEqual(expected, out.getvalue())
 
-        project.runs.filter(pipeline_name=pipeline_names[0]).update(task_exitcode=0)
+        project.runs.filter(pipeline_name=pipeline_names[0]).update(
+            task_exitcode=0, selected_groups=["group1", "group2"]
+        )
         project.runs.filter(pipeline_name=pipeline_names[1]).update(task_exitcode=1)
 
         out = StringIO()
         call_command("show-pipeline", *options, stdout=out)
         expected = (
-            " [SUCCESS] analyze_docker_image\n"
+            " [SUCCESS] analyze_docker_image (group1,group2)\n"
             " [FAILURE] analyze_root_filesystem_or_vm_image\n"
         )
         self.assertEqual(expected, out.getvalue())